requirements.txt
django==5.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openpyxl==3.0.9
email_validator==2.1.1
django-easy-select2==1.5.8
django-admin-list-filter-dropdown==1.0.3
django-tinymce==4.1.0
matplotlib==3.8.1
django-debug-toolbar==4.1.0
xlsxwriter==3.1.9